CPP03 and CPP04

+++ cercle-5/CPP01/ex01/main.cpp
@@ -6,7 +6,7 @@
 /*   By: abdoali <abdoali@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/02 18:54:21 by abdoali           #+#    #+#             */
-/*   Updated: 2026/04/16 00:53:27 by abdoali          ###   ########.fr       */
+/*   Updated: 2026/04/19 12:56:25 by abdoali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int main(void)
     int N = 5;
     Zombie *horde = zombieHorde(N, "kadir");
 
-    for (int i = 0; i < N; i++)
+    for (size_t i = 0; i < N; i++)
         horde[i].announce();
     delete[] horde;
     return (0);
