+++ cercle-5/CPP01/ex01/zombieHorde.cpp
@@ -6,7 +6,7 @@
 /*   By: abdoali <abdoali@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/02 19:06:25 by abdoali           #+#    #+#             */
-/*   Updated: 2026/04/16 00:53:27 by abdoali          ###   ########.fr       */
+/*   Updated: 2026/04/19 12:56:25 by abdoali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ Zombie* zombieHorde( int N, std::string name )
 {
     Zombie *horde = new Zombie[N];
 
-    for (int i = 0; i < N; i++)
+    for (size_t i = 0; i < N; i++)
         horde[i].setName(name);
     return horde;
 }
